Add commands to synchronize and write remote data locally

+++ src/Robo/Plugin/Commands/DrupalSyncCommands.php
@@ -43,6 +43,105 @@ class DrupalSyncCommands extends DockworkerLocalCommands {
   private $drupalRemoteSyncPodName = NULL;
 
   /**
+   * Encrypt deployed Drupal data into dockworker content archive(s).
+   *
+   * @param string $env
+   *   The deploy environment to dump from.
+   * @param string[] $opts
+   *   An array of options to pass to the builder.
+   *
+   * @option bool $no-database
+   *   Do not dump the drupal database.
+   * @option bool $no-files
+   *   Do not dump the drupal filesystem.
+   *
+   * @command repo:test-content:update
+   *
+   * @throws \Dockworker\DockworkerException
+   *
+   * @github
+   * @kubectl
+   */
+  public function encryptDrupalDatabaseFileSystemFromRemote($env, $opts = ['no-database' => FALSE, 'no-files' => FALSE]) {
+    $passphrase = $this->ask('Passphrase to encrypt content with?');
+    $this->dumpDrupalDatabaseFileSystemFromRemote($env, $opts);
+
+    exec("mkdir -p {$this->repoRoot}/data/content/");
+
+    $tmp_db_path = self::POD_TEMPORARY_FILE_LOCATION  . '/' . self::POD_DATABASE_DUMP_COMPRESSED_FILENAME;
+    $repo_db_path = "{$this->repoRoot}/data/content/db.sql.gz.gpg";
+    exec("gpg --symmetric --batch --cipher-algo AES256 --passphrase='$passphrase' $tmp_db_path");
+    exec("mv $tmp_db_path.gpg $repo_db_path");
+
+    $tmp_file_path = self::POD_TEMPORARY_FILE_LOCATION  . '/' . self::POD_FILES_DUMP_FILENAME;
+    $repo_file_path = "{$this->repoRoot}/data/content/files.tar.gz.gpg";
+    exec("gpg --symmetric --batch --cipher-algo AES256 --passphrase='$passphrase' $tmp_file_path");
+    exec("mv $tmp_file_path.gpg $repo_file_path");
+  }
+
+  /**
+   * Dumps deployed Drupal data into local archive(s).
+   *
+   * @param string $env
+   *   The deploy environment to dump from.
+   * @param string[] $opts
+   *   An array of options to pass to the builder.
+   *
+   * @option bool $no-database
+   *   Do not dump the drupal database.
+   * @option bool $no-files
+   *   Do not dump the drupal filesystem.
+   *
+   * @command deployment:content:dump
+   *
+   * @throws \Dockworker\DockworkerException
+   *
+   * @github
+   * @kubectl
+   */
+  public function dumpDrupalDatabaseFileSystemFromRemote($env, $opts = ['no-database' => FALSE, 'no-files' => FALSE]) {
+    $this->getLocalRunning();
+
+    $this->io()->title('Deployed Data Synchronization');
+
+    $this->kubernetesPodNamespace = $env;
+    $this->kubernetesSetupPods($this->instanceName, "Synchronization");
+
+    // All pods should return the same data, so simply use the first.
+    $this->drupalRemoteSyncPodName = $this->kubernetesCurPods[0];
+
+    // Determine operations to perform.
+    $this->drupalRemoteSyncDatabase = !$opts['no-database'];
+    $this->drupalRemoteSyncFiles = !$opts['no-files'];
+
+    // Dump out no-op users.
+    if (!$this->drupalRemoteSyncDatabase && !$this->drupalRemoteSyncFiles) {
+      $this->say('No operations requested. Exiting...');
+      return;
+    }
+
+    // Error Checking.
+    $this->checkRemoteDrush();
+    $this->checkLocalDrush();
+    $this->compareLocalRemoteDrupalVersions();
+
+    if ($this->drupalRemoteSyncDatabase) {
+      $this->syncDrupalDatabaseFromRemote(TRUE);
+    }
+
+    if ($this->drupalRemoteSyncFiles) {
+      $this->syncDrupalFileSystemFromRemote(TRUE);
+    }
+
+    if ($this->drupalRemoteSyncDatabase) {
+      $this->say('Database dumped to ' . self::POD_TEMPORARY_FILE_LOCATION  . '/' . self::POD_DATABASE_DUMP_COMPRESSED_FILENAME);
+    }
+    if ($this->drupalRemoteSyncFiles) {
+      $this->say('Files dumped to ' . self::POD_TEMPORARY_FILE_LOCATION  . '/' . self::POD_FILES_DUMP_FILENAME);
+    }
+  }
+
+  /**
    * Synchronizes deployed Drupal data into the local Drupal application.
    *
    * @param string $env
@@ -200,7 +299,7 @@ class DrupalSyncCommands extends DockworkerLocalCommands {
    *
    * @throws \Exception
    */
-  private function syncDrupalDatabaseFromRemote() {
+  private function syncDrupalDatabaseFromRemote($dump_only = FALSE) {
     $this->io()->newLine();
     $this->io()->section("Synchronizing Drupal database from [{$this->drupalRemoteSyncPodName}]");
     $dump_file = self::POD_TEMPORARY_FILE_LOCATION . '/' . self::POD_DATABASE_DUMP_FILENAME;
@@ -224,26 +323,28 @@ class DrupalSyncCommands extends DockworkerLocalCommands {
     $this->say("[Remote] Removing Drupal database archive file...");
     $this->runRemoteCommand('rm -f '. $gz_dump_file);
 
-    $this->say("[Docker Host] Copying Drupal database archive file to local container...");
-    $this->copyLocalFileToContainer($gz_dump_file, $gz_dump_file);
+    if (!$dump_only) {
+      $this->say("[Docker Host] Copying Drupal database archive file to local container...");
+      $this->copyLocalFileToContainer($gz_dump_file, $gz_dump_file);
 
-    $this->say("[Docker Host] Deleting Drupal database archive file...");
-    unlink($gz_dump_file);
+      $this->say("[Docker Host] Deleting Drupal database archive file...");
+      unlink($gz_dump_file);
 
-    $this->say("[Container] (optionally) Removing Drupal database archive file...");
-    $this->runLocalContainerCommand("rm -f $dump_file");
+      $this->say("[Container] (optionally) Removing Drupal database archive file...");
+      $this->runLocalContainerCommand("rm -f $dump_file");
 
-    $this->say("[Container] Decompressing Drupal database archive file...");
-    $this->runLocalContainerCommand("gunzip $gz_dump_file");
+      $this->say("[Container] Decompressing Drupal database archive file...");
+      $this->runLocalContainerCommand("gunzip $gz_dump_file");
 
-    $this->say("[Container] Importing Drupal database archive file...");
-    $this->runLocalContainerCommand("sh -c \"drush --root=/app/html sql-cli < $dump_file\"");
+      $this->say("[Container] Importing Drupal database archive file...");
+      $this->runLocalContainerCommand("sh -c \"drush --root=/app/html sql-cli < $dump_file\"");
 
-    $this->say("[Container] Removing Drupal database archive file...");
-    $this->runLocalContainerCommand("rm -f $dump_file");
+      $this->say("[Container] Removing Drupal database archive file...");
+      $this->runLocalContainerCommand("rm -f $dump_file");
 
-    $this->say("[Container] Clearing cache...");
-    $this->runLocalDrushCommand("cr");
+      $this->say("[Container] Clearing cache...");
+      $this->runLocalDrushCommand("cr");
+    }
   }
 
   /**
@@ -329,7 +430,7 @@ class DrupalSyncCommands extends DockworkerLocalCommands {
    *
    * @throws \Exception
    */
-  private function syncDrupalFileSystemFromRemote() {
+  private function syncDrupalFileSystemFromRemote($dump_only = FALSE) {
     $this->io()->newLine();
     $this->io()->section("Synchronizing Drupal filesystem from [{$this->drupalRemoteSyncPodName}]");
     $files_dump_name = self::POD_TEMPORARY_FILE_LOCATION . '/' . self::POD_FILES_DUMP_FILENAME;
@@ -349,24 +450,26 @@ class DrupalSyncCommands extends DockworkerLocalCommands {
     $this->say("[Docker Host] Copying Drupal filesystem archive to local container...");
     $this->copyLocalFileToContainer($files_dump_name, $files_dump_name);
 
-    $this->say("[Docker Host] Deleting Drupal filesystem archive...");
-    unlink($files_dump_name);
+    if (!$dump_only) {
+      $this->say("[Docker Host] Deleting Drupal filesystem archive...");
+      unlink($files_dump_name);
 
-    $this->say("[Container] Deleting container Drupal filesystem...");
-    $this->runLocalContainerCommand('rm -rf ' . self::POD_FILES_SOURCE);
+      $this->say("[Container] Deleting container Drupal filesystem...");
+      $this->runLocalContainerCommand('rm -rf ' . self::POD_FILES_SOURCE);
 
-    $this->say("[Container] Extracting remote Drupal filesystem...");
-    $this->runLocalContainerCommand("tar -xzf $files_dump_name --directory /");
+      $this->say("[Container] Extracting remote Drupal filesystem...");
+      $this->runLocalContainerCommand("tar -xzf $files_dump_name --directory /");
 
-    $this->say("[Container] Removing Drupal filesystem archive file...");
-    $this->runLocalContainerCommand('rm -f '. $files_dump_name);
+      $this->say("[Container] Removing Drupal filesystem archive file...");
+      $this->runLocalContainerCommand('rm -f ' . $files_dump_name);
 
-    $this->say("[Container] Setting overall Drupal filesystem permissions...");
-    $this->runLocalContainerCommand('/scripts/pre-init.d/70_set_drupal_tree_permissions.sh');
-    $this->say("[Container] Setting config sync permissions...");
-    $this->runLocalContainerCommand('/scripts/pre-init.d/71_set_config_sync_permissions.sh');
-    $this->say("[Container] Setting public filesystem permissions...");
-    $this->runLocalContainerCommand('/scripts/pre-init.d/71_set_public_file_permissions.sh');
+      $this->say("[Container] Setting overall Drupal filesystem permissions...");
+      $this->runLocalContainerCommand('/scripts/pre-init.d/70_set_drupal_tree_permissions.sh');
+      $this->say("[Container] Setting config sync permissions...");
+      $this->runLocalContainerCommand('/scripts/pre-init.d/71_set_config_sync_permissions.sh');
+      $this->say("[Container] Setting public filesystem permissions...");
+      $this->runLocalContainerCommand('/scripts/pre-init.d/71_set_public_file_permissions.sh');
+    }
   }
 
   /**
